Fix healers
