+++ game/src/main/resources/game.properties
@@ -93,7 +93,7 @@ world.start.gear=true
 world.npcs.collision=true
 
 # Whether NPCs can move randomly
-world.npcs.randomWalk=true
+world.npcs.randomWalk=false
 
 # Whether Lumbridge is an NPC aggression-free area
 world.npcs.safeZone=true
